EVSRESTAPI-449 Move R4 to own folder, created R5 folder, created classes from R4 examples, added R5 dependency
