+++ src/main/java/gov/nih/nci/evs/api/fhir/FHIRServerResponseException.java
@@ -9,9 +9,9 @@ public class FHIRServerResponseException extends BaseServerResponseException {
   /**
    * Instantiates a new FHIR server response exception.
    *
-   * @param theStatusCode the the status code
-   * @param theMessage the the message
-   * @param theBaseOperationOutcome the the base operation outcome
+   * @param theStatusCode the status code
+   * @param theMessage the message
+   * @param theBaseOperationOutcome the base operation outcome
    */
   public FHIRServerResponseException(
       final int theStatusCode,
@@ -23,9 +23,9 @@ public class FHIRServerResponseException extends BaseServerResponseException {
   /**
    * Instantiates a new FHIR server response exception.
    *
-   * @param theStatusCode the the status code
-   * @param theMessage the the message
-   * @param theBaseOperationOutcome the the base operation outcome
+   * @param theStatusCode the status code
+   * @param theMessage the message
+   * @param theBaseOperationOutcome the base operation outcome
    * @param e the e
    */
   public FHIRServerResponseException(
