+++ src/main/java/gov/nih/nci/evs/api/fhir/R5/ConceptMapProviderR5.java
@@ -1,0 +1,4 @@
+package gov.nih.nci.evs.api.fhir.R5;
+
+public class ConceptMapProviderR5 {
+}
