+++ src/main/java/gov/nih/nci/evs/api/fhir/R5/FhirTerminologyCapabilitiesR5.java
@@ -1,0 +1,4 @@
+package gov.nih.nci.evs.api.fhir.R5;
+
+public class FhirTerminologyCapabilitiesR5 {
+}
