+++ src/main/java/gov/nih/nci/evs/api/fhir/R5/HapiR5RestfulServlet.java
@@ -1,0 +1,4 @@
+package gov.nih.nci.evs.api.fhir.R5;
+
+public class HapiR5RestfulServlet {
+}
